CON-4792 Check sourceId when remove duplicated ordernumber Compare sourceIds of duplicated articles/variant. Delete duplicated only when sourceIds are different, otherwise it's just an update.

+++ Components/ProductToShop.php
@@ -184,7 +184,9 @@ class ProductToShop implements ProductToShopBase
         if (!empty($product->sku)) {
             $number = 'SC-' . $product->shopId . '-' . $product->sku;
             $duplicatedDetail = $this->helper->getDetailByNumber($number);
-            if ($duplicatedDetail) {
+            if ($duplicatedDetail
+                && $this->helper->getConnectAttributeByModel($duplicatedDetail)->getSourceId() != $product->sourceId
+            ) {
                 $this->deleteDetail($duplicatedDetail);
             }
         } else {
